+++ src/main/java/leets/leenk/domain/media/domain/application/dto/response/FeedMediaResponse.java
@@ -1,4 +1,4 @@
-package leets.leenk.domain.media.domain.application;
+package leets.leenk.domain.media.domain.application.dto.response;
 
 
 import leets.leenk.domain.media.domain.entity.enums.MediaType;
